feat: change average to total
