+++ src/i18n/zh-TW.json
@@ -41,8 +41,8 @@
     "pageTitle": "需量反應 一 投標",
     "modeGreen": "綠能交易",
     "modeDemand": "需量反應",
-    "avgPrice": "平均交易價",
-    "avgVolume": "平均交易量",
+    "totalPrice": "需量反應總額",
+    "totalVolume": "需量反應總量",
     "currParticipants": "即時交易人數",
     "date": "日期",
     "time": "時段",
